requirements.txt
streamlit
groq
edge-tts
python-dotenv
nest_asyncio